requirements.txt
numpy==1.25.2
Pillow==9.4.0
Pillow==10.2.0
Requests==2.31.0
scikit_learn==1.2.2
torch==2.0.1
torchvision==0.15.2a0
transformers==4.32.1
